+++ xmake/core/project/option.lua
@@ -114,7 +114,7 @@ function option:_cx_check()
             end
         end
     end
-        
+
     -- ok
     return true
 end
@@ -258,11 +258,11 @@ function option:set(name_or_info, ...)
     if type(name_or_info) == "string" then
         local args = ...
         if args ~= nil then
-            self._INFO[name_or_info] = table.unique({...})
+            self._INFO[name_or_info] = table.unique(table.join(...))
         else
             self._INFO[name_or_info] = nil
         end
-    elseif type(name_or_info) == "table" and #name_or_info == 0 then
+    elseif table.is_dictionary(name_or_info) then
         for name, info in pairs(table.join(name_or_info, ...)) do
             self:set(name, info)
         end
@@ -273,8 +273,8 @@ end
 function option:add(name_or_info, ...)
     if type(name_or_info) == "string" then
         local info = table.wrap(self._INFO[name_or_info])
-        self._INFO[name_or_info] = table.unique(table.append(info, ...))
-    elseif type(name_or_info) == "table" and #name_or_info == 0 then
+        self._INFO[name_or_info] = table.unique(table.join(info, ...))
+    elseif table.is_dictionary(name_or_info) then
         for name, info in pairs(table.join(name_or_info, ...)) do
             self:add(name, info)
         end