+++ xmake/core/project/target.lua
@@ -77,11 +77,11 @@ function target:set(name_or_info, ...)
     if type(name_or_info) == "string" then
         local args = ...
         if args ~= nil then
-            self._INFO[name_or_info] = table.unique({...})
+            self._INFO[name_or_info] = table.unique(table.join(...))
         else
             self._INFO[name_or_info] = nil
         end
-    elseif type(name_or_info) == "table" and #name_or_info == 0 then
+    elseif table.is_dictionary(name_or_info) then
         for name, info in pairs(table.join(name_or_info, ...)) do
             self:set(name, info)
         end
@@ -92,8 +92,8 @@ end
 function target:add(name_or_info, ...)
     if type(name_or_info) == "string" then
         local info = table.wrap(self._INFO[name_or_info])
-        self._INFO[name_or_info] = table.unique(table.append(info, ...))
-    elseif type(name_or_info) == "table" and #name_or_info == 0 then
+        self._INFO[name_or_info] = table.unique(table.join(info, ...))
+    elseif table.is_dictionary(name_or_info) then
         for name, info in pairs(table.join(name_or_info, ...)) do
             self:add(name, info)
         end